backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sse-starlette==3.0.3
clerk-backend-api==3.3.1
anthropic>=0.77.0
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic>=2.11.2
pydantic-settings>=2.6.0
python-json-logger==2.0.7
httpx==0.28.1
aiohttp==3.9.1
celery[redis]==5.3.6
redis>=4.7.0

# Database
sqlalchemy==2.0.23
psycopg[binary]==3.1.18  # PostgreSQL driver bundled with libpq to avoid 'libpq library not found'
alembic>=1.12.1  # Database migrations (optional, for future)
pgvector==0.3.6  # PostgreSQL vector extension for embeddings

# Document parsing
PyMuPDF==1.23.8  # fitz - for PDF text extraction
# google-cloud-documentai==3.7.0  # Google Document AI client
# google-cloud-storage==3.4.1

azure-ai-documentintelligence==1.0.0

# Embeddings & Vector Search
sentence-transformers==3.3.1  # For local embeddings (all-MiniLM-L6-v2) and cross-encoders (re-ranking)
openai==1.59.5  # For OpenAI embeddings (optional, if enabled in config)
tiktoken==0.8.0  # Token counting for compression and budget management
llmlingua==0.2.2  # Prompt compression for re-ranker token limits
Jinja2==3.1.4  # Template rendering for workflow prompts
jsonschema==4.23.0  # Strict JSON Schema validation for workflow outputs
python-docx==1.2.0
pypandoc==1.16.2
markdown==3.4.4
pandas==2.2.3
xlsxwriter==3.1.2
openpyxl==3.1.5  # Excel template reading/writing with formula preservation
prometheus-client==0.18.0
prometheus-fastapi-instrumentator>=7.1.0
boto3==1.35.54
pytz==2024.2
debugpy==1.8.19  # For VS Code debugging